requirements.txt
google-generativeai
python-dotenv
pydantic
pydantic-settings
termcolor
